[PATCH] x86-64 merge for 2.6.3
Bring the x86-64 port up to date. Lots of smaller bug fixes that have accumulated.
Also fixes another nasty bug introduced by the IA32e changes that causes BUGs at
boot for some people.

Only changes x86-64 specific files. There are some other changes that I'm sending
separately.

 - Some cleanup in NMI watchdog code
 - Fix HyperThreading CPU setup race (Suresh B. Siddha)
 - Update defconfig
 - Add a comment on why iommu_fullflush is disabled.
 - Export sys_ioctl again
 - Fix build with IA32_EMULATION=y and SYSVIPC=n
 - Remove noisy boot printks in the mptable scan.
 - Implement automatic NMI watchdog switching for real now
 - Remove redundant 32bit ioctl handlers for autofs
 - Remove CONFIG ifdefs around rtc 32bit ioctl handlers
 - Remove useless nfsctl ifdef in syscall.c (Al Viro)
 - Increase padding for prefetchw alternative
 - Check for NX bit early before setting up memory maps (Suresh B. Siddha)
 - Change Intel IA32e config description and fix help texts (Jun Nakajima)
 - Fix microcode driver build really now (Dave Jones)
 - Add nohpet option to disable HPET timer
 - Fix double semicolon in aperture.c
 - Add cmpxchg16b cpuid entry
 - Fix return value of read_pci_config_16 (Paul Menage)
 - Fix __KERNEL_COMPAT32_CS (Zachary Amsden)
 - Disable the infamous 30 minutes check in CMOS time setting
 - Update URLs in Kconfig (Petri T. Koistinen)
 - Fix ACPI interrupt source parsing for Nforce3 (Maciej W. Rozycki)
 - Fix 32bit ipc version parsing.
 - Run local APIC NMI watchdog only once a second (or less often on idle boxes)
 - Merge ACPI APIC SCI functions from i386
 - Add i8254 timer suspend code from i386
 - Merge with 2.6.2-rc3 + minor changes from i386
 - Fix empty_zero_page declaration (Greg Johnson)
 - Readd sysctls for exception/page fault trace and vsyscall32
 - Fix WCHAN
 - Fix STACK_TOP usage. Stack for 64bit processes should be at the
   top of memory now again. Also set it correctly for LINUX32_3GB.
 - Add warning fixes for gcc 3.4 and -Wdeclaration-after-statement

+++ include/asm-x86_64/pci-direct.h
@@ -28,7 +28,7 @@ static inline u8 read_pci_config_byte(u8 bus, u8 slot, u8 func, u8 offset)
 	return v;
 }
 
-static inline u8 read_pci_config_16(u8 bus, u8 slot, u8 func, u8 offset)
+static inline u16 read_pci_config_16(u8 bus, u8 slot, u8 func, u8 offset)
 {
 	u16 v; 
 	outl(0x80000000 | (bus<<16) | (slot<<11) | (func<<8) | offset, 0xcf8);
